requirements.txt
annotated-types
pydantic~=2.10.3
pydantic-core
typing-extensions
pytest~=8.3.4
gitpython~=3.1.44
pytest-cov
